requirement.txt
Django==1.11.25
django-cors-headers==3.0.2
djangorestframework==3.9.4
